Improved rules abstraction
