+++ core/development/logic/rules/abstract/Rule.ts
@@ -1,0 +1,7 @@
+export interface Rule {
+  description: string;
+  isRevealed: boolean;
+  condition: () => boolean;
+  onTrigger: () => void;
+  trigger: () => void;
+}
